pyproject.toml
[tool.poetry]
name = "to-do-app"
version = "1.0.0"
description = "A simple to-do app"
authors = []

[tool.poetry.dependencies]
Flask="1.1.1"
Jinja2="2.11.2"
python = "^3.8.2"
python-dotenv="0.12.0"
requests="2.22.0"
gunicorn = "^20.0.4"
pymongo = {version = "^3.11.2", extras = ["srv"]}

[tool.poetry.dev-dependencies]
freezegun="0.3.15"
pytest="5.4.3"
selenium="3.141.0"